rename: 패키지 이름 model -> domain으로 변경
