+++ src/main/java/domain/DefaultLottoWinningPrizeStrategy.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.HashMap;
 import java.util.Map;
