+++ src/main/java/domain/GenerateStrategy.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Set;
