+++ src/main/java/domain/LottoGame.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.LinkedHashMap;
 import java.util.List;
