+++ src/main/java/domain/LottoTicket.java
@@ -1,7 +1,6 @@
-package model;
+package domain;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
