+++ src/main/java/domain/WinningTicket.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.List;
 import java.util.Set;
