+++ src/test/java/domain/DefaultLottoWinningPrizeStrategyTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
