+++ src/test/java/domain/LottoNumberGenerateStrategyTest.java
@@ -1,9 +1,8 @@
-package model;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
