+++ src/test/java/domain/LottoTicketsTest.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
