+++ src/test/java/domain/WinningTicketTest.java
@@ -1,10 +1,9 @@
-package model;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
